% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muSignAl.R
\name{muSignAl}
\alias{muSignAl}
\title{muSignAl: Multiple Signature Algorithm.}
\usage{
muSignAl(data, feature, outcome, nrun, thres)
}
\description{
This function evaluates all the possible signatures
It requires an excel input for features as well as outcome variable
Any values not present will be imputed
}
